refactor: fix protocol architecture and implement PR review feedback
BREAKING CHANGES:
- Move AuthenticationServiceProtocol from domain to API layer (client-defined)
- Rename AuthenticationServiceImpl → AuthenticationService (remove Impl suffix)
- Extract domain models to separate models.py file
- Remove duplicate UserRepository protocol

Protocol Architecture Fixes:
- AuthenticationServiceProtocol now in /api/ (client defines interface)
- AuthenticationService in /application/ (provider implements)
- Domain models separated from protocol definitions
- Follow "Accept interfaces, return structs" principle

Code Improvements:
- Update tests to use Mock(spec=Protocol) instead of manual mocks
- Remove unnecessary workflow and documentation files
- Simplify authentication setup guide for LLM consumption
- Sync tach dependencies for new module structure

Files Changed:
- auth/api/auth_service_protocol.py (NEW - client-defined protocol)
- auth/application/authentication_service.py (RENAMED from _impl)
- auth/domain/models.py (NEW - extracted domain models)
- Remove duplicate user repository protocol
- Update all imports and type annotations
