+++ backend/tach.toml
@@ -3,9 +3,17 @@ exclude = []
 source_roots = [
     ".",
 ]
-root_module = "allow"
 
-# Infrastructure layer - can depend on all domain interfaces
+[[modules]]
+path = "<root>"
+depends_on = [
+    "infrastructure",
+    "modules.auth.api",
+    "modules.product.domain",
+    "modules.user.domain",
+    "modules.user.infrastructure",
+]
+
 [[modules]]
 path = "infrastructure"
 depends_on = [
@@ -16,18 +24,8 @@ depends_on = [
     "modules.user.domain",
 ]
 
-# Root application - can depend on infrastructure and domain
 [[modules]]
-path = "<root>"
-depends_on = [
-    "infrastructure",
-    "modules.product.domain",
-    "modules.user.domain",
-]
-
-# AI Content module - domain is pure, infrastructure can depend on shared infrastructure
-[[modules]]
-path = "modules.ai_content.domain"
+path = "modules.ai_content.api"
 depends_on = []
 
 [[modules]]
@@ -35,33 +33,50 @@ path = "modules.ai_content.application"
 depends_on = []
 
 [[modules]]
+path = "modules.ai_content.domain"
+depends_on = []
+
+[[modules]]
 path = "modules.ai_content.infrastructure"
-depends_on = [ "infrastructure"]
-
-[[modules]]
-path = "modules.ai_content.api"
-depends_on = []
-
-# Auth module - domain is pure, infrastructure can depend on shared infrastructure
-[[modules]]
-path = "modules.auth.domain"
-depends_on = []
-
-[[modules]]
-path = "modules.auth.application"
-depends_on = []
-
-[[modules]]
-path = "modules.auth.infrastructure"
-depends_on = []
+depends_on = [
+    "infrastructure",
+]
 
 [[modules]]
 path = "modules.auth.api"
-depends_on = []
+depends_on = [
+    "infrastructure",
+    "modules.auth.application",
+    "modules.auth.domain",
+    "modules.auth.infrastructure",
+    "modules.user.domain",
+    "modules.user.infrastructure",
+]
 
-# Communications module - domain is pure, infrastructure can depend on shared infrastructure
 [[modules]]
-path = "modules.communications.domain"
+path = "modules.auth.application"
+depends_on = [
+    "modules.auth.domain",
+    "modules.auth.infrastructure",
+    "modules.user.domain",
+]
+
+[[modules]]
+path = "modules.auth.domain"
+depends_on = [
+    "modules.user.domain",
+]
+
+[[modules]]
+path = "modules.auth.infrastructure"
+depends_on = [
+    "infrastructure",
+    "modules.auth.api",
+    "modules.auth.domain",
+]
+
+[[modules]]
+path = "modules.communications.api"
 depends_on = []
 
 [[modules]]
@@ -69,16 +84,15 @@ path = "modules.communications.application"
 depends_on = []
 
 [[modules]]
+path = "modules.communications.domain"
+depends_on = []
+
+[[modules]]
 path = "modules.communications.infrastructure"
 depends_on = []
 
 [[modules]]
-path = "modules.communications.api"
-depends_on = []
-
-# MercadoLibre module - domain is pure, infrastructure can depend on shared infrastructure
-[[modules]]
-path = "modules.mercadolibre.domain"
+path = "modules.mercadolibre.api"
 depends_on = []
 
 [[modules]]
@@ -86,16 +100,17 @@ path = "modules.mercadolibre.application"
 depends_on = []
 
 [[modules]]
-path = "modules.mercadolibre.infrastructure"
-depends_on = [ "infrastructure"]
-
-[[modules]]
-path = "modules.mercadolibre.api"
+path = "modules.mercadolibre.domain"
 depends_on = []
 
-# Product module - domain is pure, infrastructure can depend on shared infrastructure
 [[modules]]
-path = "modules.product.domain"
+path = "modules.mercadolibre.infrastructure"
+depends_on = [
+    "infrastructure",
+]
+
+[[modules]]
+path = "modules.product.api"
 depends_on = []
 
 [[modules]]
@@ -103,16 +118,17 @@ path = "modules.product.application"
 depends_on = []
 
 [[modules]]
-path = "modules.product.infrastructure"
-depends_on = [ "infrastructure"]
-
-[[modules]]
-path = "modules.product.api"
+path = "modules.product.domain"
 depends_on = []
 
-# User module - domain is pure, infrastructure can depend on shared infrastructure
 [[modules]]
-path = "modules.user.domain"
+path = "modules.product.infrastructure"
+depends_on = [
+    "infrastructure",
+]
+
+[[modules]]
+path = "modules.user.api"
 depends_on = []
 
 [[modules]]
@@ -120,9 +136,12 @@ path = "modules.user.application"
 depends_on = []
 
 [[modules]]
-path = "modules.user.infrastructure"
-depends_on = [ "infrastructure"]
+path = "modules.user.domain"
+depends_on = []
 
 [[modules]]
-path = "modules.user.api"
-depends_on = []
+path = "modules.user.infrastructure"
+depends_on = [
+    "infrastructure",
+    "modules.user.domain",
+]
